migration to v6
