+++ grafana-plugin/.config/webpack/webpack.config.ts
@@ -239,7 +239,7 @@ const config = async (env): Promise<Configuration> => {
 
   if (isWSL()) {
     baseConfig.watchOptions = {
-      poll: 3000,
+      // poll: 3000,
       ignored: /node_modules/,
     };
   }
