+++ grafana-plugin/src/pages/integration/Integration.tsx
@@ -1,28 +1,15 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 
 import { LabelTag } from '@grafana/labels';
-import {
-  Button,
-  HorizontalGroup,
-  VerticalGroup,
-  Icon,
-  LoadingPlaceholder,
-  IconButton,
-  ConfirmModal,
-  Drawer,
-  Alert,
-} from '@grafana/ui';
+import { Button, HorizontalGroup, VerticalGroup, LoadingPlaceholder, IconButton, Drawer, Alert } from '@grafana/ui';
 import cn from 'classnames/bind';
 import { get } from 'lodash-es';
 import { observer } from 'mobx-react';
 import moment from 'moment-timezone';
-import CopyToClipboard from 'react-copy-to-clipboard';
 import Emoji from 'react-emoji-render';
-import { RouteComponentProps, useHistory, withRouter } from 'react-router-dom';
 
 import { getTemplatesForEdit } from 'components/AlertTemplates/AlertTemplatesForm.config';
 import { TemplateForEdit } from 'components/AlertTemplates/CommonAlertTemplatesForm.config';
-import { HamburgerContextMenu } from 'components/HamburgerContextMenu/HamburgerContextMenu';
 import {
   IntegrationCollapsibleTreeView,
   IntegrationCollapsibleItem,
@@ -30,7 +17,6 @@ import {
 import { IntegrationContactPoint } from 'components/IntegrationContactPoint/IntegrationContactPoint';
 import { IntegrationHowToConnect } from 'components/IntegrationHowToConnect/IntegrationHowToConnect';
 import { IntegrationLogoWithTitle } from 'components/IntegrationLogo/IntegrationLogoWithTitle';
-import { IntegrationSendDemoAlertModal } from 'components/IntegrationSendDemoAlertModal/IntegrationSendDemoAlertModal';
 import { IntegrationBlock } from 'components/Integrations/IntegrationBlock';
 import { IntegrationTag } from 'components/Integrations/IntegrationTag';
 import { PageErrorHandlingWrapper, PageBaseState } from 'components/PageErrorHandlingWrapper/PageErrorHandlingWrapper';
@@ -42,14 +28,8 @@ import { TooltipBadge } from 'components/TooltipBadge/TooltipBadge';
 import { EditRegexpRouteTemplateModal } from 'containers/EditRegexpRouteTemplateModal/EditRegexpRouteTemplateModal';
 import { CollapsedIntegrationRouteDisplay } from 'containers/IntegrationContainers/CollapsedIntegrationRouteDisplay/CollapsedIntegrationRouteDisplay';
 import { ExpandedIntegrationRouteDisplay } from 'containers/IntegrationContainers/ExpandedIntegrationRouteDisplay/ExpandedIntegrationRouteDisplay';
-import { IntegrationHeartbeatForm } from 'containers/IntegrationContainers/IntegrationHeartbeatForm/IntegrationHeartbeatForm';
 import { IntegrationTemplateList } from 'containers/IntegrationContainers/IntegrationTemplatesList';
-import { IntegrationFormContainer } from 'containers/IntegrationForm/IntegrationFormContainer';
-import { IntegrationLabelsForm } from 'containers/IntegrationLabelsForm/IntegrationLabelsForm';
 import { IntegrationTemplate } from 'containers/IntegrationTemplate/IntegrationTemplate';
-import { MaintenanceForm } from 'containers/MaintenanceForm/MaintenanceForm';
-import { CompleteServiceNowModal } from 'containers/ServiceNowConfigDrawer/CompleteServiceNowConfigModal';
-import { ServiceNowConfigDrawer } from 'containers/ServiceNowConfigDrawer/ServiceNowConfigDrawer';
 import { TeamName } from 'containers/TeamName/TeamName';
 import { UserDisplayWithAvatar } from 'containers/UserDisplay/UserDisplayWithAvatar';
 import { WithPermissionControlTooltip } from 'containers/WithPermissionControl/WithPermissionControlTooltip';
@@ -67,22 +47,31 @@ import { useStore } from 'state/useStore';
 import { withMobXProviderContext } from 'state/withStore';
 import { LocationHelper } from 'utils/LocationHelper';
 import { UserActions } from 'utils/authorization/authorization';
-import { GENERIC_ERROR, INTEGRATION_SERVICENOW, PLUGIN_ROOT } from 'utils/consts';
-import { withDrawer } from 'utils/hoc';
-import { useDrawer } from 'utils/hooks';
+import { INTEGRATION_SERVICENOW, PLUGIN_ROOT } from 'utils/consts';
+import { PropsWithRouter, withDrawer, withRouter } from 'utils/hoc';
 import { getItem, setItem } from 'utils/localStorage';
 import { sanitize } from 'utils/sanitize';
 import { openNotification, openErrorNotification } from 'utils/utils';
 
 import { OutgoingTab } from './OutgoingTab/OutgoingTab';
+import { IntegrationActions } from './IntegrationActions';
+import { GrafanaTheme2 } from '@grafana/data';
 
 const cx = cn.bind(styles);
+
+export type IntegrationDrawerKey = typeof INTEGRATION_SERVICENOW | 'completeConfig';
+
+interface RouteProps {
+  id: string;
+}
 
 interface IntegrationProps
   extends WithDrawerConfig<IntegrationDrawerKey>,
     WithStoreProps,
     PageProps,
-    RouteComponentProps<{ id: string }> {}
+    PropsWithRouter<RouteProps> {
+  theme: GrafanaTheme2;
+}
 
 interface IntegrationState extends PageBaseState {
   isLoading: boolean;
@@ -137,10 +126,11 @@ class _IntegrationPage extends React.Component<IntegrationProps, IntegrationStat
       channelFilterIdForEdit,
       isTemplateSettingsOpen,
     } = this.state;
+
     const {
       store,
       query,
-      match: {
+      router: {
         params: { id },
       },
       drawerConfig,
@@ -583,10 +573,12 @@ class _IntegrationPage extends React.Component<IntegrationProps, IntegrationStat
   };
 
   handleAddNewRoute = () => {
-    const { alertReceiveChannelStore } = this.props.store;
     const {
-      params: { id },
-    } = this.props.match;
+      store: { alertReceiveChannelStore },
+      router: {
+        params: { id },
+      },
+    } = this.props;
 
     this.setState(
       {
@@ -621,7 +613,7 @@ class _IntegrationPage extends React.Component<IntegrationProps, IntegrationStat
   renderRoutesFn = (): IntegrationCollapsibleItem[] => {
     const {
       store: { alertReceiveChannelStore },
-      match: {
+      router: {
         params: { id },
       },
     } = this.props;
@@ -689,10 +681,12 @@ class _IntegrationPage extends React.Component<IntegrationProps, IntegrationStat
     channelFilterId: ChannelFilter['id'],
     filteringTermType?: number
   ) => {
-    const { alertReceiveChannelStore, escalationPolicyStore } = this.props.store;
     const {
-      params: { id },
-    } = this.props.match;
+      store: { alertReceiveChannelStore, escalationPolicyStore },
+      router: {
+        params: { id },
+      },
+    } = this.props;
 
     try {
       const channelFilter: ChannelFilter = await alertReceiveChannelStore.saveChannelFilter(channelFilterId, {
@@ -716,7 +710,7 @@ class _IntegrationPage extends React.Component<IntegrationProps, IntegrationStat
   onUpdateTemplatesCallback = async (data) => {
     const {
       store,
-      match: {
+      router: {
         params: { id },
       },
     } = this.props;
@@ -756,17 +750,20 @@ class _IntegrationPage extends React.Component<IntegrationProps, IntegrationStat
   };
 
   onRemovalFn = async (id: ApiSchemas['AlertReceiveChannel']['id']) => {
+    const {
+      router: { navigate },
+    } = this.props;
     await AlertReceiveChannelHelper.deleteAlertReceiveChannel(id);
-    this.props.history.push(`${PLUGIN_ROOT}/integrations/`);
+    navigate(`${PLUGIN_ROOT}/integrations/`);
   };
 
   async loadData() {
     const {
       store: { alertReceiveChannelStore, msteamsChannelStore, hasFeature },
-      match: {
+      router: {
+        navigate,
         params: { id },
       },
-      history,
     } = this.props;
 
     const promises: Array<Promise<void | { [key: string]: { alerts_count: number; alert_groups_count: number } }>> = [];
@@ -797,7 +794,7 @@ class _IntegrationPage extends React.Component<IntegrationProps, IntegrationStat
     } catch (_err) {
       if (!alertReceiveChannelStore.items[id]) {
         // failed fetching the integration (most likely it's not existent)
-        history.push(`${PLUGIN_ROOT}/integrations`);
+        navigate(`${PLUGIN_ROOT}/integrations`);
       }
     } finally {
       this.setState({ isLoading: false });
@@ -815,344 +812,6 @@ class _IntegrationPage extends React.Component<IntegrationProps, IntegrationStat
     return Promise.resolve();
   }
 }
-
-interface IntegrationActionsProps {
-  isLegacyIntegration: boolean;
-  alertReceiveChannel: ApiSchemas['AlertReceiveChannel'];
-  changeIsTemplateSettingsOpen: () => void;
-  drawerConfig: ReturnType<typeof useDrawer<IntegrationDrawerKey>>;
-}
-
-type IntegrationDrawerKey = typeof INTEGRATION_SERVICENOW | 'completeConfig';
-
-const IntegrationActions: React.FC<IntegrationActionsProps> = ({
-  alertReceiveChannel,
-  isLegacyIntegration,
-  changeIsTemplateSettingsOpen,
-  drawerConfig,
-}) => {
-  const store = useStore();
-  const { alertReceiveChannelStore } = store;
-
-  const history = useHistory();
-
-  const [confirmModal, setConfirmModal] = useState<{
-    isOpen: boolean;
-    title: any;
-    dismissText: string;
-    confirmText: string;
-    body?: React.ReactNode;
-    description?: string;
-    confirmationText?: string;
-    onConfirm: () => void;
-  }>(undefined);
-
-  const [isCompleteServiceNowConfigOpen, setIsCompleteServiceNowConfigOpen] = useState(false);
-  const [isIntegrationSettingsOpen, setIsIntegrationSettingsOpen] = useState(false);
-  const [isLabelsFormOpen, setLabelsFormOpen] = useState(false);
-  const [isHeartbeatFormOpen, setIsHeartbeatFormOpen] = useState(false);
-  const [isDemoModalOpen, setIsDemoModalOpen] = useState(false);
-  const [maintenanceData, setMaintenanceData] = useState<{
-    alert_receive_channel_id: ApiSchemas['AlertReceiveChannel']['id'];
-  }>(undefined);
-
-  const { closeDrawer, openDrawer, getIsDrawerOpened } = drawerConfig;
-
-  const { id } = alertReceiveChannel;
-
-  useEffect(() => {
-    /* ServiceNow Only */
-    openServiceNowCompleteConfigurationDrawer();
-  }, []);
-
-  return (
-    <>
-      {confirmModal && (
-        <ConfirmModal
-          isOpen={confirmModal.isOpen}
-          title={confirmModal.title}
-          confirmText={confirmModal.confirmText}
-          dismissText="Cancel"
-          body={confirmModal.body}
-          description={confirmModal.description}
-          confirmationText={confirmModal.confirmationText}
-          onConfirm={confirmModal.onConfirm}
-          onDismiss={() => setConfirmModal(undefined)}
-        />
-      )}
-
-      {alertReceiveChannel.demo_alert_enabled && (
-        <IntegrationSendDemoAlertModal
-          alertReceiveChannel={alertReceiveChannel}
-          isOpen={isDemoModalOpen}
-          onHideOrCancel={() => setIsDemoModalOpen(false)}
-        />
-      )}
-
-      {getIsDrawerOpened(INTEGRATION_SERVICENOW) && <ServiceNowConfigDrawer onHide={closeDrawer} />}
-
-      {isCompleteServiceNowConfigOpen && (
-        <CompleteServiceNowModal onHide={() => setIsCompleteServiceNowConfigOpen(false)} />
-      )}
-
-      {isIntegrationSettingsOpen && (
-        <IntegrationFormContainer
-          isTableView={false}
-          onHide={() => setIsIntegrationSettingsOpen(false)}
-          onSubmit={async () => {
-            await alertReceiveChannelStore.fetchItemById(alertReceiveChannel.id);
-          }}
-          id={alertReceiveChannel['id']}
-          navigateToAlertGroupLabels={(_id: ApiSchemas['AlertReceiveChannel']['id']) => {
-            setIsIntegrationSettingsOpen(false);
-            setLabelsFormOpen(true);
-          }}
-        />
-      )}
-
-      {isLabelsFormOpen && (
-        <IntegrationLabelsForm
-          onHide={() => {
-            setLabelsFormOpen(false);
-          }}
-          onSubmit={() => alertReceiveChannelStore.fetchItemById(alertReceiveChannel.id)}
-          id={alertReceiveChannel['id']}
-          onOpenIntegrationSettings={() => {
-            setIsIntegrationSettingsOpen(true);
-          }}
-        />
-      )}
-
-      {isHeartbeatFormOpen && (
-        <IntegrationHeartbeatForm
-          alertReceveChannelId={alertReceiveChannel['id']}
-          onClose={() => setIsHeartbeatFormOpen(false)}
-        />
-      )}
-
-      {maintenanceData && (
-        <MaintenanceForm
-          initialData={maintenanceData}
-          onUpdate={() => alertReceiveChannelStore.fetchItemById(alertReceiveChannel.id)}
-          onHide={() => setMaintenanceData(undefined)}
-        />
-      )}
-
-      <div className={cx('integration__actions')}>
-        <WithPermissionControlTooltip userAction={UserActions.IntegrationsTest}>
-          <Button
-            variant="secondary"
-            size="md"
-            onClick={() => setIsDemoModalOpen(true)}
-            data-testid="send-demo-alert"
-            disabled={!alertReceiveChannel.demo_alert_enabled}
-            tooltip={alertReceiveChannel.demo_alert_enabled ? '' : 'Demo Alerts are not enabled for this integration'}
-          >
-            Send demo alert
-          </Button>
-        </WithPermissionControlTooltip>
-
-        <div data-testid="integration-settings-context-menu-wrapper">
-          <HamburgerContextMenu
-            items={[
-              {
-                onClick: openIntegrationSettings,
-                label: 'Integration Settings',
-              },
-              {
-                label: 'ServiceNow configuration',
-                hidden: !getIsBidirectionalIntegration(alertReceiveChannel),
-                onClick: () => openDrawer(INTEGRATION_SERVICENOW),
-              },
-              {
-                onClick: openLabelsForm,
-                hidden: !store.hasFeature(AppFeature.Labels),
-                label: 'Alert group labeling',
-                requiredPermission: UserActions.IntegrationsWrite,
-              },
-              {
-                onClick: () => setIsHeartbeatFormOpen(true),
-                hidden: !showHeartbeatSettings(),
-                label: <div data-testid="integration-heartbeat-settings">Heartbeat Settings</div>,
-                requiredPermission: UserActions.IntegrationsWrite,
-              },
-              {
-                onClick: openStartMaintenance,
-                hidden: Boolean(alertReceiveChannel.maintenance_till),
-                label: 'Start Maintenance',
-                requiredPermission: UserActions.MaintenanceWrite,
-              },
-              {
-                onClick: changeIsTemplateSettingsOpen,
-                label: 'Edit Templates',
-                requiredPermission: UserActions.MaintenanceWrite,
-              },
-              {
-                onClick: () => {
-                  setConfirmModal({
-                    isOpen: true,
-                    confirmText: 'Stop',
-                    dismissText: 'Cancel',
-                    onConfirm: onStopMaintenance,
-                    title: 'Stop Maintenance',
-                    body: (
-                      <Text type="primary">
-                        Are you sure you want to stop the maintenance for{' '}
-                        <Emoji text={alertReceiveChannel.verbal_name} /> ?
-                      </Text>
-                    ),
-                  });
-                },
-                hidden: !alertReceiveChannel.maintenance_till,
-                label: 'Stop Maintenance',
-                requiredPermission: UserActions.MaintenanceWrite,
-              },
-              {
-                onClick: () =>
-                  setConfirmModal({
-                    isOpen: true,
-                    title: 'Migrate Integration?',
-                    body: (
-                      <VerticalGroup spacing="lg">
-                        <Text type="primary">
-                          Are you sure you want to migrate <Emoji text={alertReceiveChannel.verbal_name} /> ?
-                        </Text>
-
-                        <VerticalGroup spacing="xs">
-                          <Text type="secondary">- Integration internal behaviour will be changed</Text>
-                          <Text type="secondary">
-                            - Integration URL will stay the same, so no need to change {getMigrationDisplayName()}{' '}
-                            configuration
-                          </Text>
-                          <Text type="secondary">- Integration templates will be reset to suit the new payload</Text>
-                          <Text type="secondary">- It is needed to adjust routes manually to the new payload</Text>
-                        </VerticalGroup>
-                      </VerticalGroup>
-                    ),
-                    onConfirm: onIntegrationMigrate,
-                    dismissText: 'Cancel',
-                    confirmText: 'Migrate',
-                  }),
-                hidden: !isLegacyIntegration,
-                label: 'Migrate',
-                requiredPermission: UserActions.IntegrationsWrite,
-              },
-              {
-                label: (
-                  <CopyToClipboard
-                    text={alertReceiveChannel.id}
-                    onCopy={() => openNotification('Integration ID is copied')}
-                  >
-                    <div>
-                      <HorizontalGroup spacing={'xs'}>
-                        <Icon name="copy" />
-                        <Text type="primary">UID: {alertReceiveChannel.id}</Text>
-                      </HorizontalGroup>
-                    </div>
-                  </CopyToClipboard>
-                ),
-              },
-              {
-                onClick: () => {
-                  setConfirmModal({
-                    isOpen: true,
-                    title: 'Delete Integration?',
-                    body: (
-                      <Text type="primary">
-                        Are you sure you want to delete <Emoji text={alertReceiveChannel.verbal_name} /> ?
-                      </Text>
-                    ),
-                    onConfirm: deleteIntegration,
-                    dismissText: 'Cancel',
-                    confirmText: 'Delete',
-                  });
-                },
-                hidden: !alertReceiveChannel.allow_delete,
-                label: (
-                  <Text type="danger">
-                    <HorizontalGroup spacing={'xs'}>
-                      <Icon name="trash-alt" />
-                      <span>Delete Integration</span>
-                    </HorizontalGroup>
-                  </Text>
-                ),
-                requiredPermission: UserActions.IntegrationsWrite,
-              },
-            ]}
-          />
-        </div>
-      </div>
-    </>
-  );
-
-  function openServiceNowCompleteConfigurationDrawer() {
-    const isServiceNow = getIsBidirectionalIntegration(alertReceiveChannel);
-    const isConfigured = alertReceiveChannel.additional_settings?.is_configured;
-    if (isServiceNow && !isConfigured) {
-      setIsCompleteServiceNowConfigOpen(true);
-    }
-  }
-
-  function getMigrationDisplayName() {
-    const name = alertReceiveChannel.integration.toLowerCase().replace('legacy_', '');
-    switch (name) {
-      case 'grafana_alerting':
-        return 'Grafana Alerting';
-      case 'alertmanager':
-      default:
-        return 'AlertManager';
-    }
-  }
-
-  async function onIntegrationMigrate() {
-    try {
-      await AlertReceiveChannelHelper.migrateChannel(alertReceiveChannel.id);
-      setConfirmModal(undefined);
-      openNotification('Integration has been successfully migrated.');
-      await Promise.all([
-        alertReceiveChannelStore.fetchItemById(alertReceiveChannel.id),
-        alertReceiveChannelStore.fetchTemplates(alertReceiveChannel.id),
-      ]);
-    } catch (_err) {
-      openErrorNotification(GENERIC_ERROR);
-    }
-  }
-
-  function showHeartbeatSettings() {
-    return alertReceiveChannel.is_available_for_integration_heartbeat;
-  }
-
-  async function deleteIntegration() {
-    try {
-      await AlertReceiveChannelHelper.deleteAlertReceiveChannel(alertReceiveChannel.id);
-      history.push(`${PLUGIN_ROOT}/integrations`);
-      openNotification('Integration has been succesfully deleted.');
-    } catch (_err) {
-      openErrorNotification(GENERIC_ERROR);
-    }
-  }
-
-  function openIntegrationSettings() {
-    setIsIntegrationSettingsOpen(true);
-  }
-
-  function openLabelsForm() {
-    setLabelsFormOpen(true);
-  }
-
-  function openStartMaintenance() {
-    setMaintenanceData({ alert_receive_channel_id: alertReceiveChannel.id });
-  }
-
-  async function onStopMaintenance() {
-    setConfirmModal(undefined);
-
-    await AlertReceiveChannelHelper.stopMaintenanceMode(id);
-
-    openNotification('Maintenance has been stopped');
-    await alertReceiveChannelStore.fetchItemById(id);
-  }
-};
 
 interface IntegrationHeaderProps {
   alertReceiveChannelCounter: AlertReceiveChannelCounters;
@@ -1290,4 +949,7 @@ const IntegrationHeader: React.FC<IntegrationHeaderProps> = ({
   }
 };
 
-export const IntegrationPage = withRouter(withMobXProviderContext(withDrawer<IntegrationDrawerKey>(_IntegrationPage)));
+export const IntegrationPage = withRouter<
+  RouteProps,
+  Omit<IntegrationProps, 'store' | 'meta' | 'theme' | 'drawerConfig'>
+>(withMobXProviderContext(withDrawer<IntegrationDrawerKey>(_IntegrationPage)));
